requirements.txt
streamlit
google-generativeai
google-cloud-logging
python-dotenv
pytest
pytest-mock
google-cloud-secret-manager